Generate code stubs from the UML diagrams
Describe the relation between the code and UML in README.md.

The code generation support for Python in Umbrello is lackluster,
which is why the documentation style does not match the ReStructuredText
format at the moment. Converting to that style will be the next task.

Furthermore, the way the modules are referred to (e.g. core.AudioCatalogLayer)
is not okay; we should use relative references (e.g. from . import AudioCatalogLayer).
That is also a task for a later commit.

The enums in overview were not exported, they will also need to be added later.